fixed modul

+++ hw5/hw5/Calculator.fs
@@ -1,10 +1,11 @@
-﻿module hw5.Calculator
+﻿module hw5.Calculator 
 
     let Calculate val1 operation val2 =
-        let a = val1 |> double
-        let b = val2 |> double
+        let a = val1 |> int
+        let b = val2 |> int
         match operation with
         | "+" -> a + b
         | "-" -> a - b
         | "*" -> a * b
         | "/" -> a / b
+
